backend/pyproject.toml
[project]
name = "mp-undercover"
version = "0.1.0"
description = "WeChat Undercover Table Game"
authors = [
    {name = "xingxiaolin"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Flask>=3.0.0",
    "Flask-SQLAlchemy>=3.1.1",
    "Flask-Migrate>=4.0.5",
    "gunicorn>=21.2.0",
    "redis>=5.0.1",
    "requests>=2.31.0",
    "wechatpy>=1.8.18",
    "cryptography>=41.0.7",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",
    "pymysql>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.6",
    "fakeredis>=2.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C4"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["src"]
combine-as-imports = true
